error checks, refactoring

+++ results_archive.py
@@ -1,16 +1,16 @@
 #results_archive.py
 #archiving results files to glacier
+#http://boto3.readthedocs.io/en/latest/reference/services/glacier.html
 
 
 import boto3
 import os
-from ast import literal_eval
 import json
 import time
 
 AWS_S3_RESULTS_BUCKET = "gas-results"
 AWS_SNS_JOB_ARCHIVE_QUEUE = 'syun0_archive_jobs'
-GLACIER_ARCHIVE_DELAY_IN_SECONDS = 60
+GLACIER_ARCHIVE_DELAY_IN_SECONDS = 1800
 AWS_GLACIER_VAULT = "ucmpcs"
 AWS_S3_RESULTS_BUCKET = "gas-results"
 AWS_REGION_NAME = os.environ['AWS_REGION_NAME'] if ('AWS_REGION_NAME' in  os.environ) else "us-east-1"
@@ -32,7 +32,7 @@ def archive_job(result_file_key=None):
 
 if __name__ == '__main__':
     #connect to SQS and get the message queue
-    sqs = boto3.resource('sqs', region_name='us-east-1')
+    sqs = boto3.resource('sqs', region_name=AWS_REGION_NAME)
     queue = sqs.get_queue_by_name(QueueName=AWS_SNS_JOB_ARCHIVE_QUEUE)
 
     while True:
@@ -40,46 +40,59 @@ if __name__ == '__main__':
         if(messages):
             for message in messages:
                 #read message body
-                body = literal_eval(message.body)
-                data = literal_eval(body.get('Message'))
-                job_id = data.get('job_id_archive')
-                result_file_key = data.get('result_file_key')
-                completion_time = data.get('completion_time')
+                body = json.loads(message.body)
+                data = json.loads(body['Message'])
+                job_id = data['job_id_archive']
+                result_file_key = data['result_file_key']
+                completion_time = data['completion_time']
 
                 current_time = int(time.time())
                 time_elapsed = current_time - completion_time
                 print("elapsed time for job " + job_id + ": " + str(time_elapsed) + " seconds")
                 if time_elapsed > GLACIER_ARCHIVE_DELAY_IN_SECONDS:
-                    #archive file and get archived id
-                    results_file_archive_id = archive_job(result_file_key=result_file_key)
-
-                    #delete original file from S3
-                    s3 = boto3.client('s3')
-                    s3.delete_object(
-                        Bucket=AWS_S3_RESULTS_BUCKET,
-                        Key=result_file_key,
-                    )
-
-                    #record archived ID in dynamodb
+                    #check if user is still free (not upgraded during wait time)
+                    dynamodb = boto3.resource('dynamodb', region_name=AWS_REGION_NAME)
+                    ann_table = dynamodb.Table(AWS_DYNAMODB_ANNOTATIONS_TABLE)
                     try:
+                        response = ann_table.get_item(Key={'job_id': job_id})
+                    except Exception:
+                        print('Error: failed to retrieve job information from database')
+                    data = response.get('Item')
+                    #if user is already premium, delete message and do nothing
+                    if(data['user_role'] == 'premium_user'):
+                        print('user is premium, deleting archive job request')
+                        message.delete()
+                    #if user is still free, go on with glacier archiving
+                    else:
+                        #archive file and get archived id
+                        results_file_archive_id = archive_job(result_file_key=result_file_key)
+
+                        #delete original file from S3
+                        s3 = boto3.client('s3')
+                        s3.delete_object(
+                            Bucket=AWS_S3_RESULTS_BUCKET,
+                            Key=result_file_key,
+                        )
+
+                        #record archived ID in dynamodb
                         dynamodb = boto3.resource('dynamodb', region_name=AWS_REGION_NAME)
                         ann_table = dynamodb.Table(AWS_DYNAMODB_ANNOTATIONS_TABLE)
-                    except Exception:
-                        print("Failed to retrieve connect to database")
+                        try:
+                            ann_table.update_item(
+                                Key = {'job_id': job_id},
+                                UpdateExpression='SET results_file_archive_id = :val1, archived = :val2',
+                                ExpressionAttributeValues={
+                                    ':val1': results_file_archive_id,
+                                    ':val2': 'true'
+                                }
+                            )
+                        except Exception:
+                            print('Error: failed to update database')
 
-                    ann_table.update_item(
-                        Key = {'job_id': job_id},
-                        UpdateExpression='SET results_file_archive_id = :val1, archived = :val2',
-                        ExpressionAttributeValues={
-                            ':val1': results_file_archive_id,
-                            ':val2': 'true'
-                        }
-                    )
+                        print("job " + job_id + " archived to Glacier")
 
-                    print("job " + job_id + " archived to Glacier")
-
-                    #delete the message
-                    message.delete()
+                        #delete the message
+                        message.delete()
 
         else:
             print("No notifications found")
